+++ src/settings/util/index.js
@@ -1,2 +1,3 @@
 export * from './convertData';
+export * from './filteringConditionsDataOptions';
 export * from './route';