+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add ability for duplicate set. Refs UIOAIPMH-8.
 * Rename sets endpoints. Refs UIOAIPMH-16.
 * Add details view for set. Refs UIOAIPMH-10.
+* Add filtering conditions for create/edit set. Refs UIOAIPMH-6.
 
 ## [1.2.0](https://github.com/folio-org/ui-oai-pmh/tree/v1.2.0) (2020-07-03)
 [Full Changelog](https://github.com/folio-org/ui-oai-pmh/compare/v1.1.0...v1.2.0)